wfw-search-filters/src/app/search-filters-module/services/refinement.service.spec.ts
import { Refinement } from './../models/refinement';
import { WfwEventsService } from 'wfw-shared/services/wfw-events.service';
import { AuthService } from 'wfw-shared/services/auth.service';
import { HttpModule } from '@angular/http';
import { ApiClientService } from 'wfw-shared/services/api-client.service';
import { SearchResultsService } from 'wfw-shared/services/search-results.service';
import { TreeDataTransformService } from './tree-data-transform.service';
import { browser } from 'protractor';
import { Component, OnInit } from '@angular/core';
import { async, ComponentFixture, TestBed, inject } from '@angular/core/testing';
import { TreeModule, TreeNode } from 'primeng/primeng';
import { RefinementService } from './refinement.service';
import { AvailableRefinementsComponent } from './../components/available-refinements/available-refinements.component';
import { SelectedRefinementsComponent } from './../components/selected-refinements/selected-refinements.component';
import { SearchFiltersComponent } from './../components/search-filters/search-filters.component';

describe('RefinementService', () => {
  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ HttpModule],
      providers: [RefinementService, TreeDataTransformService, SearchResultsService,
        ApiClientService, AuthService, WfwEventsService]
    });
  });

  it('refinement service should be created', inject([RefinementService], (service: RefinementService) => {
    expect(service).toBeTruthy();
  }));

  it('refinement service should set applied refinements', inject([RefinementService], (service: RefinementService) => {
    service.getAppliedRefinements().subscribe( (data) => {
      expect(data).toEqual( [{ TreeNode: { data: 'data', parent: { data: 'Data Type' } } }] );
    });
    spyOn(service, 'getDataTypeData');
    service.setAppliedRefinements([{data: 'data' , parent: {data: 'Data Type' }}]);
    expect(service.getDataTypeData).toHaveBeenCalledWith('data');
   }));

   it('refinement service should splice the array', inject([RefinementService], (service: RefinementService) => {
    const arr = [{key: 'A'}, {key: 'B'}];
    const objToBeSpliced = {key: 'B'};
    const resArr = [{key: 'A'}];
    const retArr = service.spliceOutNodeFromSelectedFileArray(objToBeSpliced, arr);
    expect(retArr).toEqual(resArr);
   }));

   it('refinement service should find unapplied refinements', inject([RefinementService], (service: RefinementService) => {
    service.getUnAppliedRefinements().subscribe( (data) => {
      expect(data).toEqual( { TreeNode: ''} );
    });
   }));

   it('should remove partial selection', inject([RefinementService],
    (service: RefinementService) => {
      const param1 = {partialSelected: true};
      const param2 = {parent: {partialSelected: true}};
      const ret = {};
      const rslt1 = service.removePartialSelection(param1);
      expect(rslt1).toEqual(ret);
      const rslt2 = service.removePartialSelection(param2);
      expect(rslt2).toEqual(ret);
   }));

   it('should clean up selected file array', inject([RefinementService],
    (service: RefinementService) => {
      const param = [{parent: {parent: {}}}];
      const ret = [ { parent: {  } } ];
      const rslt = service.cleanUpSelectedFileArray(param);
      expect(rslt).toEqual(ret);
   }));

   it('should splice out node from selected file array', inject([RefinementService],
    (service: RefinementService) => {
      const param = [{key: 'one'}, {key: 'two'}];
      const param2 = {key: 'two'};
      const ret = [{key: 'one'}];
      const rslt = service.spliceOutNodeFromSelectedFileArray(param2, param);
      expect(rslt).toEqual(ret);
   }));

   it('should identify child node in selected file array', inject([RefinementService],
    (service: RefinementService) => {
      const param = [{key: 'one'}, {key: 'two'}];
      const param2 = {key: 'two'};
      const param3 = {key: 'three'};
      const rslt = service.childNodeInSelectedFile(param2, param);
      const rslt2 = service.childNodeInSelectedFile(param3, param);
      expect(rslt2).toEqual(false);
   }));

   // Commented currently as this fails, working on resolution
   xit('refinement service should call parentPartialSelectionUnselection', inject([RefinementService], (service: RefinementService) => {
    const curPrnt = {label: 'Data Sources', data: 'Data Sources',
      children: [{label: 'EDM (1)', data: 'EDM', children: undefined, expanded: true,
      parent: [{label: 'Data Sources', data: 'Data Sources', children:
      [{label: 'EDM (1)', data: 'EDM', children: undefined, expanded: true, parent: undefined,
      partialSelected: true}, {}], expanded: true, parent: undefined, partialSelected: true}],
      partialSelected: true}, {}], expanded: true, parent: undefined, partialSelected: true};
    const selectdFle = [];
    const rslt = service.parentPartialSelectionUnselection(curPrnt, selectdFle);
    expect(rslt).toBe([]);
   }));
  });
